# Improve Rational performance